scripts/data_collection/requirements.txt
pandas
requests
spotipy
nest-asyncio
aiohttp
python-dotenv

